requirements.txt
django==1.9
djangorestframework==3.4
django-filter==0.15
pygments
requests
bs4
